Fixed %post script error during upgrade
During post script we initiates vdsm-tool set-saslpasswd call. On
upgrade we have mixed code that may lead to an error while calling
vdsm-tool script, as in the related bug.
In this patch we move the call to set-saslpasswd to init scripts. In
pre-script phase we can be sure that the newer code is running.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1168689
Change-Id: Ib7808f4f217ba3d6d30a60fab6233e36a29f0098
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35744
Reviewed-by: Dan Kenigsberg <danken@redhat.com>