iscsiadm.iface_list() to return iface.net_ifacename
Currently iface_list() only returns a list with iface names. In order to
properly instantiate the corresponding iface class, this patch adds to
iface_list() all parameters returned by the command 'iscsiadm -m iface' in
a tuple of namedtuples, each one representing a distinct iface.

The issue that pushed this code was iface class not being instantiated
with proper iface.net_ifacename. This patch is using the new
iface_list() format to catch iface.net_ifacename and pass it to iface
class instatiation.

Change-Id: I6d0192860b53e24524f58245c92d45f8146bc725
Signed-off-by: Amador Pahim <apahim@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35807
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>