Changing persisted value to libvirt logrotate configure
Setting persisted to false means that is-configured won't check the
status of this file. This to avoid differences between ovirt-3.4
configuration to ovirt-3.5 in /etc/logrotate.d/libvirtd

In ovirt-3.4 we set the prefix of the configuration without the conf
version, therefore in 3.5 we alerted that libvirt is not configured
and asked to run vdsm-tool configure all-over again without having any
differences between the versions.

Change-Id: I3f1c4d5d513e8699543727fff9f30a8213c66bfd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1168695
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35758
Reviewed-by: Dan Kenigsberg <danken@redhat.com>