Revert "supervdsm: Add zombiereaper to supervdsm"
Introduction of zombiereaper is causing communication failures between
vdsm and supervdsm. The change register signal handler which causes
multiprocessing to fail.

There was python bug opened to fix multiprocessing issue but it was not
backported to 2.6.

It is also not included on el7's python-2.7.5 so we are reverting on master as well.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1168217
Related-To: https://bugzilla.redhat.com/1181624
Change-Id: I90a76ad689938cc36fd9adbf73bd16e9a408cbb5
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/36961
Reviewed-by: Dan Kenigsberg <danken@redhat.com>