hostdev: report totalvfs
Totalvfs is a parameter defined by hardware that tells us how many
virtual functions we can spawn (if bandwidth of the bus allows it).
This patch adds totalvfs field to host device tree to device params.
The value is read from /sys/bus/pci/devices/x/sriov_totalvfs where x
is name of the device in the format of nnnn:nn:nn.n.

Change-Id: Ic54228deb64db200615ef8e2ba8f51e449e68022
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35891
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>