jsonrpc: Fix callMethod in JsonRpcClient
In order to call a verb using JsonRpcClient we have several methods that
we can use. We can call method asynchronosuly (call_async) or
synchronously (call). callMethod help with building JsonRpcRequest and
raises an Error in case of any issues occuring on the other side.
callMethod was using call_batch which is not needed so we have decided
to merge both methods and get rid of call_batch as well as fix issues
around single vs batch.

Change-Id: I7cf78f3b7f5d32c10339bbe6b49ce9b93662b4bb
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/35935
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>