stomp: Always pop the frame for frame handler
We do not need to check whether _frame_handler has attribute handle_frame
because we have only two frame handlers. One is used when connecting and the
other used when serving the content.

Following classes are used use frame handlers:
- AsyncClient (client)
- StompAdapterImpl (server)

Both of them provide handle_frame. We could keep this check if we plan to
provide more implementations but currently there is no need to have more.

Change-Id: I9cd3c5b9ad70cf49e1b105260548720d5b34ea9a
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/35940
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>