profiling: add support for memory profiling
There are many python memory profilers, but most of them are
tailored for interactive usage, so additional integration work
is needed to use them in a daemon, like VDSM.

This patch adds support for dowser:
<http://www.aminus.net/wiki/Dowser>
<https://pypi.python.org/pypi/dowser/0.2>

which has the following benefits which makes it in a better position
to be used with VDSM:
+ self contained, includes WEB UI
+ allows live monitoring
+ tailored for cherrypy usage, so friendly towards server applications.
+ easily portable (pure package)

It must be noted that dowser has some drawbacks as well
- significat dependencies, both in number and in size (with respect
  to VDSM standards): cherrypy, PIL
- based on gc module

As per the cpu profile already added, this code is meant to be
used only as debug aid or in development environments.
The feature is controlled by a config tunable and disabled by default;
the feature disable itself if any dependency is missed.

Change-Id: Ib56b65513e0118b68cd43791bf655c928d6a26e2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/32019
Reviewed-by: Dan Kenigsberg <danken@redhat.com>