vm: ensure graphics configuration data in recovery
Graphics Devices may be not sent by Engine. To cover that VDSM
reconstructs them from other VM configuration data - which is safe to do
since the information is there, just organized in a different way.

The recovery path skips that step, so the configuration is inconsistent.
Later in the creation path the configuration is assumed to be present
(either fixed by VDSM or given by Engine), but it is not, and this makes
the domain initialization fail in the last stages.

This is evident when upgrading Hosted Engine.
This issue was unnoticed so far because the reccomended upgrade path
is to migrate VMs away from the host being upgraded, and that avoid and
triggers the existing safety nets for the missing configuration data.

Change-Id: I6ed99b3e1a62d1e4c5bff5722bce9a966b193c73
Bug-Url: https://bugzilla.redhat.com/1168709
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36021
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>