handling of network events for multipe connection
When a host is SPM there are 2 connections open. When we disconnect
domain irs connection is closed. This command generates 'Client closed'
event which is wrongly processed for both connections.

In order to fix it we need to have a way to distinguish multipe
connections and notify correct on on 'Client closed' event. All the
other events should be handled as before.

Change-Id: Iaf37c9e3fff8b244a0eeb30c63a6932703948806
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1172757