ip route: parse only first part of each route
This patch makes sure that the output of `ip route show table all` has
one line corresponding to exactly one route. It then ignores whatever
follows the first backslash of a route text. The motivation is to ignore
'cache' lines, which are of no interest to Vdsm at the moment.

Change-Id: I5b655ddbbb2fdaa7ec57bc460bc593e40bac977e
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36184
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Tested-by: Ondřej Svoboda <osvoboda@redhat.com>