spec: Require newer selinux-policy-targeted on EL7
In EL7 and fedora, selinux did not allow sanlock to send signals to
vdsm. When the master domain is not accessible on the spm, sanlock was
trying to terminate and kill vdsm, but since it is not allowed to send
signals to vdsm, the attempt would fail and the watchdog would reboot
the machine.

The new policy fixes this issue, and now vdsm should terminate cleanly
or will be killed without rebooting the machine.

We need also a package for Fedora, but the package was not released yet.

Change-Id: I99ce144a289a4e15ac85331062ad6d5e2588bd46
Bug-Url: https://bugzilla.redhat.com/1141658
Relates-to: https://bugzilla.redhat.com/1172147
Relates-to: https://bugzilla.redhat.com/1142454
Relates-to: https://bugzilla.redhat.com/1152594
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36188
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36226
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>