hostdev: add dynamic udev rule creation for iommu groups
Detaching a device from a host (via detach_detachable) spawns
a new vfio endpoint in /dev/vfio/$name, where $name = iommu group
number. In order to start a VM with this endpoint assigned (= vm
with device passed through) we need rw permissions for QEMU.

This patch adds dynamic udev rule creation/deletion for when device is
detached/reattached from or to the host.

Creating an udev rule will happen one time for each IOMMU group, that
is passed through to the host. That means we're expecting the call to
only happen on VM creation, where the number of such VMs is very limited
(possibly 1 or few).

Change-Id: Ieac8d58e01d7277e535a2101d522961816ea88eb
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36268
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>