gluster: fix volume name parsing in getVmVolumeInfo
It is permitted to use slashes '/' in gluster mount strings even
though they're prohibited in the volume name.
In order to recognize and strip them out we should use getRealPath
instead of getRemotePath that translates slashes into permitted
underscores.

It is now allowed to mount gluster volumes with slashes (serv:/volume)
and correctly identifying the volume name and its information.

Change-Id: Icfdbc573ec2aec31b78323253f8178425a07302c
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/36237
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Tested-by: Sandro Bonazzola <sbonazzo@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>