storageServer: Fix connection error handling
We used to log a traceback and then raise the original error instance
(with a truncated traceback), which would cause another useless
traceback for the same error.

Now we keep the original exception info and re-raising it, so it is not
masked by errors raised during cleanup. Since we raise useful exception,
we don't need the additional log; having 2 tracebacks for same error is
not helping anyone.

Change-Id: I746cd71d065849dda7a2b53d7b74b53265a36b36
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36331
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>