Add tests for getAllVolumes
Profiling show that FileStorageDomain.getAllVolumes is too slow, taking
more than 30% of cpu time when starting or recoverting 20 vms with 3000
disks on NFS storage domain. Each call in this setup took 44 seconds of
cpu time.

This patch adds tests documenting the somewhat surprizing behavior of
getAllVolumes to make sure we do not break it in the next patch.

Change-Id: Ia43d8970ffc9bb8ee7679232ff3ee95cb2cda36c
Relates-To: https://bugzilla.redhat.com/1177634
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36589
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>