numa: virt: remove one more minidom user
The NUMA support code relies on libvirt's internal XML representation
to build its state and to track VCPU placement.

A side effect of this already questionable usage is that VDSM has
to parse a large XML document -larger than a VM domain definition,
since the private representation contains libvirt's internal
bookkeping data each time it is asked for NUMA stats.

While we look for a more robust and more efficient way
to reconstruct this information, this patch switches
to a more efficient XML processing to offer some short-term
quick help.

Change-Id: Ice97563fc11be79cc3215db434761024b3e244dc
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36626
Reviewed-by: Dan Kenigsberg <danken@redhat.com>