init: wait for network service
The network service is not synchronous. It returns before all of the
interfaces that it started are actually up.

This is particularly problematic when upgrading from ovirt-3.4 to
ovirt-3.5, as one of the first pre-run tasks is to upgrade network
configuration to the "unified persistence" model.

This patch waits up to 10 seconds for vdsm-controlled interfaces to
obtain their IPv4 address. It's a hack, an ugly hack, that does not solve
the inherent race. The race would be solved only when we tap into
netlink's event monitor and stall ifup until the interface is actually
up.

Change-Id: I89abe7a83673c2232f8b249f23a39e27e01e50be
Bug-Url: https://bugzilla.redhat.com/1174611
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36647
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36703
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>