models: make IpConfig deal with blank configuration from NetDevice gracefully
This patch is a part of a bigger change striving to remove IpConfig's
namedtuple which merely adds an unnecessary layer in network configurators
and to introduce a boolean 'dhcpv4' parameter in setupNetworks (to mirror
the 'dhcpv6' parameter) which will deprecate the legacy 'bootproto' parameter.

As the first milestone IpConfig now gains the ability to initialize itself
with no parameters so it is possible to gradually replace the usage of the
namedtuple (IpConfig.ipConfig) in configurators (and in models.NetDevice)
with IpConfig itself (NetDevice.ip, soon to be renamed to NetDevice.ipconfig).

Next, the @property NetDevice.ipConfig, which just wraps information stored
in IpConfig (and its members IPv4 and IPv6) can be removed.

Finally, IpConfig will move from the string 'bootproto' to a boolean 'dhcpv4'.

I think it is supported to configure (if a user pleases) a network with no IP
configuration. Hence, I removed the ConfigNetworkError exception ("You need to
specify IPv4 or IPv6 or both address").

Change-Id: I09a5d1e3468004f529c2bc9708cff2243c16241a
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36712
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Tested-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>