fileUtils: Convert unhelpful warnings to debug messages
When a directory exists, fileUtils.createdir() logs a warning, although
it has no context to tell if the existing directory is expected in the
normal flow, or an unexpected error condition.

This warning causes users to worry about non-existing issues and feel
less confident about the product. It also makes developer life harder,
since searching for warnings results in many unrelated messages.

This patch changes the warning to debug message, which may be helpful
for debugging.  Code that wants to warn about existing directories
should not call fileUtils.createdir() but the underlying os.makedirs()
and handle the EEXIST error.

Change-Id: I7800b860eb81334b63abefcca9a21a552331458f
Bug-Url: https://bugzilla.redhat.com/1129587
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/36764
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI