fileUtils: Fix bug when mode is not specified
We used to modify the directory bit on the mode parameter, which is
ignored by os.makedirs(), so the mode directory bit is set. This ensure
that when comparing existing path mode, we detect the error where a file
exists in this path, instead of a directory.

However, if the mode parameter was not specified, we did not check the
file type when we get EEXIST error, so the function succeeded instead of
failing loudly:

    with open("file", "w"):
        createdir("file")

Replacing the smart code with explicit check for directory bit and
permissions bits eliminate this bug and is more clear. Now we raise
OSError with ENOTDIR errno in this case.

Change-Id: I669791fb09ddd18f9b2e6320cf98839ed07aa831
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/36781
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Continuous-Integration: Jenkins CI