vm: cleaner retrieval of graphics stats
Device object and device configuration
data are kept in sync, so there is no need to scan all the
device configuration data to report it to Engine.

Instead, just iterate on the Graphics Device classes.
This new code is also a bit faster in the microbenchmarks, but
since this method is already very low on profile, total
performance gain is negligible.

Change-Id: I337e2baa5a349c2e750f883c4ca697aaafc55bf3
Related-To: https://bugzilla.redhat.com/1177634
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36791
Reviewed-by: Dan Kenigsberg <danken@redhat.com>