stomp: Make sure the \0 is read before slicing the buffer
In python if you try and slice past the end of the list\string you get
an empty string and not an IndexError

The original code check that we got all the data and sliced past the
ending \0. But on rare cases where the message ended exactly with all
the content but the \0 was not received.

The slicing code would silently ignore the fact that the index doesn't
exist and would return an empty string and switch the parser state
making all future data have the wrong offset.

This in turn made the command '\0SEND' instead of 'SEND' which has no
handler and where silently ignored. The only way to solve this was to
reconnect.

Change-Id: Ib6d7d25719f814b97ebd2cc4c33ff25dc32e1357
Bug-Url: http://bugzilla.redhat.com/1181233
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-On: http://gerrit.ovirt.org/37018
Reviewed-on: http://gerrit.ovirt.org/36806
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Oved Ourfali <oourfali@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>