tests: deduplicate clientIF fakes
both clientIFTests and vmfakelib modules
had an implementation of clientIF fake.

Some tests could have specific needs for Fakes,
and it may have sense to have specific Fakes
tied to specific tests, when one size does not
fits all (fake.VM is rapidly going this route).

However, duplication of code is always lurking,
so there is should be a clean and not debatable
case to have specific (= duplicate) Fakes.

It doesn't seem the case for faking ClientIF yet,
so this patch reunite the two implementations
into one, to be placed into vmfakelib.

Change-Id: I4b27c3d37f62a1e19bf252290fd6816f5c65adcc
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36893
Reviewed-by: Dan Kenigsberg <danken@redhat.com>