net config: make sure ifcfg files are not persisted on node
When setting up networking, a multitude of ifcfg files may be modified.
If we are using the (default) 'unified' persistence mode, we expect that
none of these modification is to be persisted for the next boot.

However already-persisted ifcfg files would; unpersisted files won't,
which lead to the inconsistent state reported in the bug.

In this patch we make sure that none of our changes to ifcfg files is to
be persisted.

Change-Id: Ia61b2385ce0bf44a67f627caf5b73b68992349fd
Bug-Url: https://bugzilla.redhat.com/1165203
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36985
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>