networkTests: remove a broken workaround for NM interfering in DHCP tests
When introducing testDhclientLeases which tested DHCPv4 only I let the test
suite add an IPv4 address to the client side of a veth pair (on the other
there is dnsmasq serving as a DHCP server) so NetworkManager will not run
its own dhclient there. I also added this workaround to testSetupNetworks-
AddDelDhcp).

There is a more appropriate way to unmanaging devices, and that is by passing
NM_CONTROLLED=no in an ifcfg file to nmcli conn load, although this workaround
is probably specific to distributions that use ifcfg network configuration
(as an Arch user I may address this later, but patches are welcome).

Ever since I added IPv6 capability to the dnsmasq wrapper used in both tests
they require that NM be stopped or told not to manage veths because it sets
the sysctl property disable_ipv6 on all devices out of security concerns, see
e.g. https://bugzilla.redhat.com/show_bug.cgi?id=1102064

If NM is running the two tests are skipped with the message added recently:

  SKIP : Failed to add the IPv6 address fdb3:84e5:4ff4:55e3::1/64 to device
  veth_xrSWAagFYW: ['RTNETLINK answers: Permission denied']; NetworkManager
  may have set the sysctl disable_ipv6 flag on the device, please see e.g.
  RH BZ #1102064

At least Fedora 21 and EL7 are affected, EL6 is not because it relies on the
'network' initscript. The NM versions on Fedora 21 and EL7 are currently:

  0.9.9.1-35.git20140326.4dba720.el7 and
  0.9.10.1-1.4.20150115git.fc21

Let's remove the imperfect approach now.

In the following patches we will inform NetworkManager to "unmanage" both
veth sides (and later also dummies) by passing an ifcfg file with
NM_CONTROLLED=no in it.

Change-Id: Idba14753bf9cd37ec1659a49ba7e13b9478f3913
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37041
Reviewed-by: Dan Kenigsberg <danken@redhat.com>