protocoldetector: Fix race in protocoldetector tests
To test removal of stale connections, we were waiting 2 cleanup
intervals on the client side, to make sure that the server would wake up
and remove the our stale connection.

However, since the server runs in a separate thread, and the cleanup
timeout was initialized when the server was started, it was possible
that the server did not wake up yet when the client woke up and try to
send a message. This caused test_reject_very_slow_client_concurrency()
to fail from time to time.

Now we initialize _next_cleanup when the object is created, so we know
exactly when the server should wake up, eliminiating this race.

Change-Id: Ic51f502172ccaf960dd1d9d33f7410d23f7fc491
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37046
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>