vm: properly implement readPauseCode
When restarted, VDSM needs to know if a VM was paused for
disk space exhausted, in order to properly extend the disk(s)
before to (re)start it again.

Since the move to libvirt, this code was not implemented.
However, a couple of libvirt APIs which somehow are gone
unnoticed so far, make it possible to implement it.

This patch rebuilds readPauseCode on top of these APIs.
Currently, readPauseCode reports these codes:

NOERR  - if everything was fine.
ENOSPC - if one or more disks was filled.
EOTHER - if one or more disks exprienced any other error.

Change-Id: Ie2f200e2856669bb7dbcc904898f326b939b1ddd
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37118
Reviewed-by: Dan Kenigsberg <danken@redhat.com>