zombiereaper: Raise if used incorrectly
zombiereaper was trying to reap pid when it was not initialized
properly. This lead to slient failures and and leaking zombies processes
for the process lifetime.

In commit 0636b8607b (Revert "supervdsm: Add zombiereaper to supervdsm")
we removed zombiereaper from supervdsm, but there are still callers that
assume it is inizialized. This patch ensures callers will fail loudly if
zombiereaper was not initialzied.

The tests are more carefull now not change the Python process state when
importing tests from the test module.

Change-Id: Ib3a275a3f022c7b686ce07693b7c0b147762b043
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37260
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>