tests: mark testRunningStats as broken
tests/vmApiTests was an early attempt to test that
the output of API verb [vdsm.virt.]VM.getStats()
was compliant to the schema.

Because of how the sampling code currently
works, to check for running stats is racy.
Moreover, much more solid schemaCompliancyTests was
recently added.

The best course of action in the medium/long term
is to refacor both the sampling and the testing
code. To have more testable code will avoid races
in the tests.

Until that time, to avoid noise and to waste time,
this patch marks the current broken test as such.

Change-Id: I314a84d1d067256a40afd1ec4523b68ea94aa70d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37275
Reviewed-by: Dan Kenigsberg <danken@redhat.com>