network: store non-Vdsm ifcfgs with unified persistence
Unified restoration first removes everything in the running config and
then tries to regenerate everything that's in the persisted config,
problem is, that original devices are not persisted.

In this patch we (temporarily) use ifcfg persistence for non-Vdsm devices
(even when unified persistence is set), so we are able to restore them.

Change-Id: Iaac0d82863334238e6db71b4a6a4f6b76a260223
Bug-Url: https://bugzilla.redhat.com/1188251
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37453
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>