vm: make _getNicStats a function
_getNicStats used to be a classmethod of VmStatsThread,
but there are no valid reasons anymore to keep it like that.

This patch makes it a plain function, to make room
for more refactoring in this area.

There are no changes in logic.
Except for moving code around, the only significant change
is the removal of the now-useless 'cls' argument.

Change-Id: Ia191fce7a1f3330013a621d97038bb32fd3cd563
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37457
Reviewed-by: Dan Kenigsberg <danken@redhat.com>