vm: stats: cache metadata value
vcpuLimit can be set only by Engine (or anyone else
impersonating it using the UpdateVmPolicy API).

It seems a little wasteful to poll, albeit infrequently
(every 15s), for changes in a value which we know
when and how it will change.

Instead of doing so, this patch adds a simple cache
variable, to be transparently updated after
each UpdateVmPolicy.

Please note that the same approach was taken for I/O
tune, based on the assumption that VDSM is the only
controller of libvirt.

Since this assumption is ingrained in the very roots
of how VDSM works, it seem fair to extend to (V)CPU
tune as well.

Change-Id: I574b66764f8c3980fdbef3bc969b483a04542704
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37592
Reviewed-by: Dan Kenigsberg <danken@redhat.com>