virt: Invalidate drive blockDev when path changes
After live storage migration between mixed domain types, a drive
blockDev must be reinitialized. We initialize this property in __init__
for cdrom or floppy devices, or lazily for other devices, and cache the
computed value for later calls.

The initialization was moved to single location, to make it easier to
understand the code.

Drive.path is now a property so modifying it after LSM invalidates
cached blockDev value.

For easier debugging, each time Drive.path changes we log the old and
new values.

Change-Id: I387dd55843701597f595eb234d30db44a93a8535
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37552
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>