virt: introduce support for periodic operations
In order to keep track of the state of a VM,
the virt code needs to perform periodic operations,
most notably statistics sampling and bookkeeping.
Currently, these operation are piggybacked
in the sampling threads, just because this is a simple
and cheap way to run some code asynchronously
and periodically.

Besides overloading 'sampling' with bookkeeping,
these operation can't fit nicely in the bulk stats
approach that we want to adopt to improve scalability.

What we really need is a reliable method to run async,
periodically operations on VM instances.
Later, we can fit sampling in this framework
in a nicer and cleaner way.

As part of the sampling improvement effort, we added
two generic utilities which provide great buliding
blocks to create a periodic operations infrastructure,
'executor' and 'scheduler'.

So, this patch lays down the basic infrastructure
to run async/periodic VM-related operations, and start
moving bookkeeping operations to them.

Change-Id: I834c067b1d3dba16a1f8c83e555d2cc04cbf32ed
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37593
Reviewed-by: Dan Kenigsberg <danken@redhat.com>