virt: Check if a drive is chuhked before extending
Previously extendDriveVolume() was checking if a drive is a block device
and assumed that callers checked the drive format. The check was unneeded
when called from extendDrivesIfNeeded(), and spreading the logic between
different methods makes the code harder to understand.

Now extendDriveVolume() requires that callers are doing the required
checks and callers are checking Drive.chunked before calling it.

When we support LSM between mixed storage domain, Drive.chunked will
encapsulate the logic for handling replication of file-based drive to
block based storage domain.

In merge(), we are always dealing with cow volumes, so checking for
blockDev is enough. Using Drive.chunked for clarity.

Change-Id: I48416dde002272d45761296f176b97f4d870849d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37573
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>