netinfo: report DHCP from configuration to avoid confusion after stopping DHCP
Until now, after moving away from reporting DHCP based on ifcfg files, VDSM
relied on dhclient's lease files. Even after a network was reconfigured to
stop using DHCP and thus dhclient was stopped, DHCP would still be reported
from the still valid leases (which preserve the state of network configuration
as obtained from a DHCP server, to rebind later to the same address if
allowed).

Although I introduced this behaviour intentionally, VDSM must stop reporting
bootproto4 (in upstream, dhcpv4) for a network that is reverted to static
configuration (which is to be effective immediately not to confuse the engine
and the user).

The currently active network configuration is queried to find out if DHCPv4
is to be reported, or not. If a network is not configured, leases are
consulted again.

testSetupNetworksAddDelDhcp reverts from DHCP-configured network to static
configuration and checks that DHCP is no longer reported (indeed, netinfo
returns the configured values for networks, now).

Change-Id: I11848101731c1d787e5b5b9c1f8fdb8a4eed6599
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37617
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1279824
Reviewed-on: https://gerrit.ovirt.org/48846
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>