netinfo: report DHCP from configuration to avoid confusion after stopping DHCP
Note: this concerns reporting of DHCP for networks only, not for individual
devices, so far. (If we require the engine to read network configuration from
the devices, typically bridges, we will have to report DHCP, as configured on
the network, even for the devices.)

Until now, after moving away from reporting DHCP based on ifcfg files, VDSM
relied on dhclient's lease files. Even after a network was reconfigured to
stop using DHCP and thus dhclient was stopped, DHCP would still be reported
from the still valid leases (which preserve the state of network configuration
as obtained from a DHCP server, to rebind later to the same address if
allowed).

Although I introduced this behaviour intentionally, VDSM must stop reporting
dhcpv4 and dhcpv6 for a network that is reverted to static configutation
(which is to be effective immediately not to confuse the engine and the user).

The currently active network configuration is queried to find out if DHCPv4/6
is to be reported, or not. If a network is not configured, leases are
consulted again.

testSetupNetworksAddDelDhcp reverts from DHCP-configured network to static
configuration and checks that DHCP is no longer reported (indeed, netinfo
returns the configured values for networks, now).

Change-Id: I11848101731c1d787e5b5b9c1f8fdb8a4eed6599
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37617
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>