_get_gateway: allow duplicated gateways if they are in fact the same routes
This happens with a DHCP libvirt network (e.g. the ovirtmgmt bridge) for which
1) dhclient(-script) adds a gateway to the main table and also
2) dhclient-script notifies VDSM's dynamic source routing thread which adds
   the same gateway to a dedicated routing table (and corresponding routing
   rules)

In this configuration the gateway acquired via DHCP(v4) was not reported so
let's start reporting it now.

Before solving the referenced bug, only gateways from the 'main' routing table
would be reported, not the ones in separate tables. Since dhclient(-script) is
not supposed to handle source routing we have to consult all tables which VDSM
itself creates.

Change-Id: Ia59155dd959e58b81e4149fa4cf2bd406f6404a1
Bug-Url: https://bugzilla.redhat.com/1186330
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37667
Reviewed-by: Dan Kenigsberg <danken@redhat.com>