vm: add missing vm id to get stats of down vm
The stats that are returned for a VM which is down does not contains the
id of the VM.

This was broken by 7f577ab94b35f2cb2560d3ae49d6b0f753588474, which failed
to put the vmId in the 'Exited Vm Stats', but only in the 'Running Vm Stats'

That causes a regression since the engine uses this id,
without it hibernated VM is not destroyed, but remains down on the host.
Thus, adding the vm-id back.

Change-Id: I9534f2105b9a32ebd4b28dae42e265094a3655b9
Bug-Url: https://bugzilla.redhat.com/1191405
Signed-off-by: Arik Hadas <ahadas@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37694
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>