schema: require vmId in {Exited,Running}VmStats
On commit bb45b4b619c9a30a8fb35865fbd99f680295ac15
we were reminded in an harsh way that vmId must
always be present in the Vm.getStats() output.

One of the problems was that the vmId field was injected
the Vm.getStats() output by API.py, and this value was
not documented in the schema.

So, this patches update the schema to reflect the
reality of the Engine's needs.

Change-Id: Id4ac677f4321277fa274510c5a0b830b09e0b1bd
Related-To: https://bugzilla.redhat.com/1191405
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37696
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>