virt: Fix limit when calculating next volume size
Drive.getNextVolumeSize() limit was wrong in two ways. First,
Drive.truesize is *not* the drive true size on block-based drive (it
is a duplicate of Drive.apparentsize). Second, Drive.truesize is using
bytes but the code assumed that it is using megabytes. The result was
that next volume size was limited only by the vg free size.

Drive.getNextVolumeSize() requires now a capacity argument.  To prevent
the confusion with mixing different size units, we use now safer naming
convention.

Vm.extendDriveVolume() requires now a capacity argument. The capacity
was already available but unused in the drive monitoring code.

When starting live storage migration or live merge, we use now
libvirt.virDomain.blockInfo() to get drive capacity for the initial
extension.

Change-Id: I4098dfc07184085e613f17b2c48d32e47888106c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37726
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>